+++ RedPandaIDE/src/parser/cpptokenizer.cpp
@@ -486,8 +486,6 @@ QString CppTokenizer::getPreprocessor()
 
 QString CppTokenizer::getWord()
 {
-    // Skip spaces
-    skipToNextToken();
     QString result;
     // Get next word...
     while (true) {
@@ -506,6 +504,8 @@ QString CppTokenizer::getWord()
         if (currentWord.isEmpty()) {
             break;
         } else {
+            //skip spaces
+            skipToNextToken();
             if (currentWord!="operator") {
                 // Skip template contents, but keep template variable types
                 if (*mCurrent == '<') {
@@ -823,6 +823,7 @@ void CppTokenizer::skipTemplateArgs()
             break;
         case '>':
             lastBracketPos = mCurrent;
+            shouldExit = true;
             break;
         }
         if (shouldExit)